% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baier_robertson.R
\name{baier_robertson}
\alias{baier_robertson}
\alias{baier_robertson,Raster,Raster-method}
\alias{baier_robertson,data.table,missing-method}
\alias{baier_robertson,character,character-method}
\title{PET calculation by Baier and Robertson method}
\usage{
baier_robertson(x, y = NULL)

\S4method{baier_robertson}{Raster,Raster}(x, y = NULL)

\S4method{baier_robertson}{data.table,missing}(x)

\S4method{baier_robertson}{character,character}(x, y = NULL)
}
\arguments{
\item{x}{Raster* object; data.table (see details); filename (character; see details)}

\item{y}{Raster* object or filename; minimum temperature data (required for Raster and character inputs)}
}
\value{
Raster* object; data.table
}
\description{
The function \code{baier_robertson} computes Potential Evapotranspiration (PET) by Baier and Robertson method.
}
\details{
If `x` is a data.table, its columns should be named: "lon", "lat", "date", "tmin", and "tmax".

If `x`, `y`, is a filename, it should point to a *.nc file.
}
\keyword{internal}
