% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega_calc.R
\name{calculate_omega}
\alias{calculate_omega}
\title{Calculate omega}
\usage{
calculate_omega(x, y)
}
\arguments{
\item{x}{a RasterBrick with delta values}

\item{y}{a RasterBrick with latitude values}
}
\value{
a RasterBrick object
}
\description{
Auxiliary function to calculate omega
}
\keyword{internal}
