% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_length.R
\name{day_length}
\alias{day_length}
\title{Day Length}
\usage{
day_length(x)
}
\arguments{
\item{x}{a character string with the path to the data file.}
}
\value{
data.table
}
\description{
Calculate Day Length for a given raster template
}
\keyword{internal}
