% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download various evapotranspiration data products}
\usage{
download_data(name = "all", project_folder = ".")
}
\arguments{
\item{name}{a character string with the name(s) of the desired data set. Suitable options are:
\itemize{
\item{"all" for all of the below listed data sets (default),}
\item{"era5" for ERA5,}
\item{"era5_land" for ERA5-Land,}
\item{"fldas" for FLDAS,} 
\item{"gldas-clsm" for GLDAS CLSM,}
\item{"gldas-noah" for GLDAS NOAH,}
\item{"gldas-vic" for GLDAS VIC,}
\item{"gleam" for GLEAM V3,}
\item{"jra_55" for JRA-55,}
\item{"merra2" for MERRA-2,}
\item{"terraclimate" for TerraClimate,}
}}

\item{project_folder}{a character string with the path where evapoRe will be hosted. Inside it the required subfolders will be created.}
}
\value{
No return value, called to download the required data sets.
}
\description{
The function \code{download_data} downloads the selected data product.
}
\examples{
\donttest{
download_data("gldas-vic", tempdir())
}
}
