% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_env_data.R
\name{download_env_data}
\alias{download_env_data}
\title{Download environmental data products}
\usage{
download_env_data(
  path = "",
  data_name = "all",
  variable = "all",
  domain = "raw",
  time_res = "monthly"
)
}
\arguments{
\item{path}{Character; local path where data will be saved.}

\item{data_name}{Character; dataset name or "all". Options:
\itemize{
\item{"cru",}
\item{"era5-land",}
\item{"mswx-past",}
\item{"fluxcom",}
\item{"terraclimate",}
\item{"all".}
}}

\item{domain}{Character; spatial domain. Options: "raw", "land".}

\item{time_res}{Character; time resolution. Currently only "monthly" supported.}
}
\value{
Invisible; downloads selected data to the specified path.
}
\description{
Downloads radiation, temperature, dew point, humidity, albedo, surface pressure,
wind speed, and related data from CRU, ERA5-Land, MSWX, FluxCom, and TerraClimate.
}
\examples{
\donttest{
download_env_data(path = tempdir(), data_name = "mswx-past", variable = "all")
download_env_data(path = tempdir(), data_name = "cru", variable = "t2m")
download_env_data(path = tempdir(), data_name = "all", variable = "all")
}
}
