% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_terraclimate.R
\name{download_terraclimate}
\alias{download_terraclimate}
\title{TerraClimate data downloader}
\usage{
download_terraclimate(
  folder_path = ".",
  domain = "raw",
  time_res = "monthly",
  variable = "e"
)
}
\arguments{
\item{folder_path}{a character string with the path where the data will be downloaded.}

\item{domain}{a character string with the desired domain data set. Suitable options are:
\itemize{
\item{"raw" for default available spatial coverage,}
\item{"global" for data sets with global (land and ocean) coverage,}
\item{"land" for data sets with land only coverage,}
\item{"ocean", for data sets with ocean only coverage.}
}}

\item{time_res}{a character string with the desired time resolution. Suitable options are:
\itemize{
\item{"monthly",}
\item{"yearly".}
}}

\item{variable}{a character string indicating the variable to download. Suitable options are:
\itemize{
\item{"e" for evapotranspiration,}
\item{"tmin" for minimum temperature,}
\item{"tmax" for maximum temperature,}
\item{"tavg" for average temperature.}
}}
}
\value{
No return value, called to download the data set.
}
\description{
Function for downloading TerraClimate data.
}
\keyword{internal}
