% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esr_dates.R
\name{esr_dates}
\alias{esr_dates}
\title{Extraterrestrial Solar Radiation (ESR) Dates}
\usage{
esr_dates(x)
}
\arguments{
\item{x}{a character string with the path to the data file.}
}
\value{
data.table
}
\description{
Fix dates for best ESR estimates
}
\keyword{internal}
