% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gldas_clsm_cz_ts.R
\docType{data}
\name{gldas_clsm_cz_ts}
\alias{gldas_clsm_cz_ts}
\title{Monthly Evapotranspiration data}
\format{
A data.table with 120 obs. of 4 variables:
\describe{
  \item{date}{IDate format \%Y-\%m-\%d}
  \item{value}{monthly average values}
  \item{name}{full name of the data set}
  \item{type}{source type of the data set}
}
}
\source{
National Aeronautics and Space Administration (NASA)
}
\usage{
gldas_clsm_cz_ts
}
\description{
A subset of GLDAS CLSM monthly Evapotranspiration data in mm over Czechia. More detail about raw data can be found \href{https://ldas.gsfc.nasa.gov/gldas}{here}.
}
\keyword{datasets}
