% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gldas_clsm_esp_ts.R
\docType{data}
\name{gldas_clsm_esp_ts}
\alias{gldas_clsm_esp_ts}
\title{Monthly Evapotranspiration data}
\format{
A data.table with 120 obs. of 2 variables:
\describe{
  \item{date}{IDate format \%Y-\%m-\%d}
  \item{value}{monthly average values}
}
}
\source{
National Aeronautics and Space Administration (NASA)
}
\usage{
gldas_clsm_esp_ts
}
\description{
A subset of GLDAS CLSM monthly Evapotranspiration data in mm over Spain. More detail about raw data can be found \href{https://ldas.gsfc.nasa.gov/gldas}{here}.
}
\keyword{datasets}
