% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hargreaves_samani.R
\name{hargreaves_samani}
\alias{hargreaves_samani}
\alias{hargreaves_samani,Raster,Raster,Raster-method}
\alias{hargreaves_samani,data.table,missing,missing-method}
\alias{hargreaves_samani,character,character,character-method}
\title{PET calculation by Hargreaves and Samani method}
\usage{
hargreaves_samani(x, y = NULL, z = NULL)

\S4method{hargreaves_samani}{Raster,Raster,Raster}(x, y = NULL, z = NULL)

\S4method{hargreaves_samani}{data.table,missing,missing}(x)

\S4method{hargreaves_samani}{character,character,character}(x, y = NULL, z = NULL)
}
\arguments{
\item{x}{Raster* object; data.table (see details); filename (character; see details)}

\item{y}{Raster* object or filename; maximum temperature data (required for Raster and character inputs)}

\item{z}{Raster* object or filename; minimum temperature data (required for Raster and character inputs)}
}
\value{
Raster* object; data.table
}
\description{
The function \code{hargreaves_samani} computes Potential Evapotranspiration (PET) by Hargreaves and Samani method.
}
\details{
If `x` is a data.table, its columns should be named: "lon", "lat", "date", "tavg", "tmin", and "tmax".

If `x`, `y`,`z` is a filename, it should point to a *.nc file.
}
\keyword{internal}
