% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_index.R
\name{heat_index}
\alias{heat_index}
\title{Annual Heat Index}
\usage{
heat_index(x)
}
\arguments{
\item{x}{a character string with the path to the data file.}
}
\value{
data.table
}
\description{
Calculate Annual Heat Index
}
\keyword{internal}
