% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jensen_haise.R
\name{jensen_haise}
\alias{jensen_haise}
\title{PET calculation by Jensen Haise method}
\usage{
jensen_haise(x)
}
\arguments{
\item{x}{a RasterBrick object having average temperature}
}
\value{
a RasterBrick object
}
\description{
Function to calculate pet
}
\keyword{internal}
