% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcguinness_bordne.R
\name{mcguinness_bordne}
\alias{mcguinness_bordne}
\title{PET calculation by McGuinness Bordne method}
\usage{
mcguinness_bordne(x)
}
\arguments{
\item{x}{a RasterBrick object having average temperature}
}
\value{
a RasterBrick object
}
\description{
Function to calculate pet
}
\keyword{internal}
