% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oudin.R
\name{oudin}
\alias{oudin}
\alias{oudin,Raster-method}
\alias{oudin,data.table-method}
\alias{oudin,character-method}
\title{PET calculation by Oudin method}
\usage{
oudin(x)

\S4method{oudin}{Raster}(x)

\S4method{oudin}{data.table}(x)

\S4method{oudin}{character}(x)
}
\arguments{
\item{x}{Raster* object; data.table (see details); filename (character; see details)}
}
\value{
Raster* object; data.table
}
\description{
The function \code{oudin} computes Potential Evapotranspiration (PET) by Oudin method
}
\details{
If `x` is a data.table, its columns should be named: "lon", "lat", "date", and "value"

If `x` is a filename, it should point to a *.nc file.
}
\keyword{internal}
