% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet_aux.R
\name{pet_aux}
\alias{pet_aux}
\title{Common math for pet calculations}
\usage{
pet_aux(x)
}
\arguments{
\item{x}{a RasterBrick object}
}
\value{
a RasterBrick object
}
\description{
Auxiliary function to calculate pet
}
\keyword{internal}
