% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet_br.R
\name{pet_br}
\alias{pet_br}
\title{PET calculation by Baier and Robertson method}
\usage{
pet_br(tmax, tmin)
}
\arguments{
\item{x}{a RasterBrick object generated by \code{\link{pet_aux}}}

\item{tavg}{a RasterBrick object having average temperature}
}
\value{
a RasterBrick object
}
\description{
Function to calculate pet
}
\examples{
#Downloading temperature data
download_terraclimate(folder_path = ".","land", variable = "tmin") 
download_terraclimate(folder_path = ".","land", variable = "tmax")
#Loading the nc files     
tmax_nc <- "terraclimate_tmax_land_19580101_20221231_025_monthly.nc"
tmin_nc <- "terraclimate_tmin_land_19580101_20221231_025_monthly.nc"
# Converting nc files to RasterBrick 
tmax_brick <- raster::brick(tmax_nc)
tmin_brick <- raster::brick(tmin_nc)
#Calculating PET 
pet_br <- pet_calc(method_name = "br", tmax = tmax_brick, tmin = tmin_brick)
}
\keyword{internal}
