% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet_calc.R
\name{pet_calc}
\alias{pet_calc}
\title{PET methods calculation}
\usage{
pet_calc(method_name, tavg, tmax, tmin)
}
\arguments{
\item{method_name}{a character string indicating the method name to calculate
pet. Available options are:
\itemize{
\item{"hs" for calculating pet by Hargreves Samani method,}
\item{"od" for calculating pet by Oudin methode,}
\item{"mb" for calculating pet by McGuinness and Bordne methode,}
\item{"jh" for calculating pet by Jensen Haise methode,}
\item{"br",for calculating pet by Baier and Robertson.}
}}

\item{tavg}{a RasterBrick object having average temperature}

\item{tmax}{a RasterBrick object having maximum temperature}

\item{tmin}{a RasterBrick object having minimum temperature}
}
\value{
a RasterBrick object
}
\description{
Function to calculate various pet methods
}
\examples{
\dontrun{
calculate pet by Hargreaves Samani method 
x <- tavg.nc  # or character string with path to data file 
y <- tmax.nc # or character string with path to data file 
z <- tmin.nc # or character string with path to data file 
x_brick <- raster::brick(x) 
y_brick <- raster::brick(y) 
z_brick <- raster::brick(z)
pet <- pet_calc(method_name = "hs",x_brick,y_brick,z_brick)

calculate pet by Oudin method 
x <- tavg.nc  # or character string with path to data file
pet <- pet_calc(method_name = "od", tavg = x_brick)
}
}
