% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet_check.R
\name{pet_check}
\alias{pet_check}
\title{PET method name checker}
\usage{
pet_check(method_name)
}
\arguments{
\item{method_name}{a character string.}
}
\value{
No return value, called to download the data set.
}
\description{
Function to check if PET method is available
}
\keyword{internal}
