% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet_hs.R
\name{pet_hs}
\alias{pet_hs}
\title{PET calculation by Hargreaves-Samani method}
\usage{
pet_hs(tavg, tmax, tmin)
}
\arguments{
\item{tavg}{a RasterBrick object having average temperature}

\item{tmax}{a RasterBrick object having maximum temperature}

\item{tmin}{a RasterBrick object having minimum temperature}
}
\value{
a RasterBrick object
}
\description{
Function to calculate pet
}
\examples{
\dontrun{
x <- tavg.nc  # or character string with path to data file 
y <- tmax.nc # or character string with path to data file 
z <- tmin.nc # or character string with path to data file 
x_brick <- raster::brick(x) 
y_brick <- raster::brick(y) 
z_brick <- raster::brick(z) 
pet <- pet_hs(tavg = x_brick, tmax = y_brick, tmin = z_brick)
}
}
\keyword{internal}
