% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet_mb.R
\name{pet_mb}
\alias{pet_mb}
\title{PET calculation by McGuinness Bordne method}
\usage{
pet_mb(tavg)
}
\arguments{
\item{tavg}{a RasterBrick object having average temperature}
}
\value{
a RasterBrick object
}
\description{
Function to calculate pet
}
\examples{
\dontrun{
x <- tavg.nc  or charater string with path to data file 
x_brick <- raster::brick(x) 
pet <- pet_mb(tavg = x_brick)
}
}
\keyword{internal}
