% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet_params.R
\name{pet_params_calc}
\alias{pet_params_calc}
\title{Potential Evapotranspiration (PET) Parameters}
\usage{
pet_params_calc(x)
}
\arguments{
\item{x}{a data.table, its columns should be named: "lon", "lat", "date", and "value".}
}
\value{
data.table
}
\description{
Calculate parameters related to PET, including `esr_date`, `omega`, `ext_rad`,
`delta`,`lat_rad`, and, `dr`
}
\keyword{internal}
