% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary.R
\name{plot_summary}
\alias{plot_summary}
\title{Evapotranspiration summary plot}
\usage{
plot_summary(x)
}
\arguments{
\item{x}{a csv generated by \code{\link{make_ts}}}
}
\value{
ggplot object
}
\description{
Function for plotting Evapotranspiration summary (line, matrix, box, and density)
}
