% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thornthwaite.R
\name{thornthwaite}
\alias{thornthwaite}
\title{PET calculation by Thornthwaite method}
\usage{
thornthwaite(x)
}
\arguments{
\item{x}{a RasterBrick object having average temperature}
}
\value{
a RasterBrick object
}
\description{
Function to calculate pet
}
\keyword{internal}
