% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turc.R
\name{turc}
\alias{turc}
\alias{turc,missing,missing,missing-method}
\alias{turc,Raster,Raster,Raster-method}
\alias{turc,character,character,character-method}
\title{Calculate Potential Evapotranspiration (PET) using the Turc Method}
\usage{
turc(tavg, rs, rh, x = NULL)

\S4method{turc}{missing,missing,missing}(x)

\S4method{turc}{Raster,Raster,Raster}(tavg, rs, rh, x = NULL)

\S4method{turc}{character,character,character}(tavg, rs, rh, x = NULL)
}
\arguments{
\item{tavg}{Raster* object or file path; average temperature (°C)}

\item{rs}{Raster* object or file path; shortwave radiation (MJ m⁻² day⁻¹)}

\item{rh}{Raster* object or file path; relative humidity (%)}

\item{x}{Optional. A `data.table` with columns: "lon", "lat", "date", "tavg", "rs", "rh"}
}
\value{
A `RasterBrick` or `data.table` with calculated PET values (mm/day).
}
\description{
The function \code{turc} implements the Turc method to estimate PET based on daily mean temperature (`tavg`),
shortwave radiation (`rs`), and relative humidity (`rh`).
}
\details{
Computes PET using the Turc empirical formula, which uses temperature, radiation, and humidity.


This function supports three input types:
\itemize{
  \item Raster inputs: provide `Raster*` objects for `tavg`, `rs`, and `rh`.
  \item File paths: provide character paths to NetCDF or raster files.
  \item data.table input: provide a `data.table` named `x` with columns \code{lon}, 
  \code{lat}, \code{date}, \code{tavg}, \code{rs}, \code{rh}.
}
}
\keyword{internal}
