% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blaney_criddle.R
\name{blaney_criddle}
\alias{blaney_criddle}
\title{PET calculation by Blaney Criddle method}
\usage{
blaney_criddle(x)
}
\arguments{
\item{x}{a RasterBrick object having average temperature}
}
\value{
a RasterBrick object
}
\description{
Calculate pet
}
\keyword{internal}
