% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esr.R
\name{esr}
\alias{esr}
\title{Extraterrestrial Solar Radiation (ESR)}
\usage{
esr(x)
}
\arguments{
\item{x}{a character string with the path to the data file.}
}
\value{
data.table
}
\description{
Calculate ESR for a given raster template
}
\keyword{internal}
