% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamon.R
\name{hamon}
\alias{hamon}
\title{PET calculation by Hamon method}
\usage{
hamon(x)
}
\arguments{
\item{x}{a RasterBrick object having average temperature}
}
\value{
a RasterBrick object
}
\description{
Function to calculate pet
}
\keyword{internal}
