% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oudin.R
\name{oudin}
\alias{oudin}
\title{PET calculation by Oudin method}
\usage{
oudin(x)
}
\arguments{
\item{x}{a RasterBrick object having average temperature}
}
\value{
a RasterBrick object
}
\description{
Calculate pet
}
\keyword{internal}
