% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oudin.R
\name{oudin}
\alias{oudin}
\title{PET calculation by Oudin method}
\usage{
oudin(x)
}
\arguments{
\item{tavg}{a RasterBrick object having average temperature}
}
\value{
a RasterBrick object
}
\description{
Function to calculate pet
}
\examples{
\dontrun{
tavg_brick <- raster::brick("terraclimate_tavg_land_19580101_20221231_025_monthly.nc")
pet_od <- pet(method = "od", tavg = tavg_brick)}
}
\keyword{internal}
