% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oudin.R
\name{oudin}
\alias{oudin}
\alias{oudin,Raster-method}
\alias{oudin,data.table-method}
\alias{oudin,character-method}
\title{Calculate Potential Evapotranspiration (PET) using Oudin method}
\usage{
oudin(x)

\S4method{oudin}{Raster}(x)

\S4method{oudin}{data.table}(x)

\S4method{oudin}{character}(x)
}
\arguments{
\item{x}{Raster* object, character file path, or data.table (see details)}
}
\value{
RasterBrick or data.table of PET values (mm/day)
}
\description{
The function \code{oudin} computes PET using the Oudin method.
}
\details{
For Raster inputs, provide a raster object or file path for average temperature.
For `data.table` input, provide a table with columns: "lon", "lat", "date", and "tavg".
}
\keyword{internal}
