% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet_br.R
\name{pet_br}
\alias{pet_br}
\title{PET calculation by Baier and Robertson method}
\usage{
pet_br(tmax, tmin)
}
\arguments{
\item{x}{a RasterBrick object generated by \code{\link{pet_aux}}}

\item{tavg}{a RasterBrick object having average temperature}
}
\value{
a RasterBrick object
}
\description{
Function to calculate pet
}
\examples{
\dontrun{
x <- tavg.nc  or character string with path to data file 
x_brick <- raster::brick(x) 
pet <- pet_br(tavg = x_brick)
}
}
\keyword{internal}
