% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet_br.R
\name{pet_br}
\alias{pet_br}
\title{PET calculation by Baier and Robertson method}
\usage{
pet_br(tmax, tmin)
}
\arguments{
\item{x}{a RasterBrick object generated by \code{\link{pet_aux}}}

\item{tavg}{a RasterBrick object having average temperature}
}
\value{
a RasterBrick object
}
\description{
Function to calculate pet
}
\examples{
\donttest{ 
download_terraclimate(folder_path = ".","land", variable = "t")} 
tmax_brick <- raster::brick("terraclimate_tmax_land_19580101_20221231_025_monthly.nc")
tmin_brick <- raster::brick("terraclimate_tmin_land_19580101_20221231_025_monthly.nc")
pet_br <- pet_calc(method = "br", tmax = tmax_brick, tmin = tmin_brick)
}
\keyword{internal}
