% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet_calc.R
\name{pet_calc}
\alias{pet_calc}
\title{Calculate PET by various offline methods}
\usage{
pet_calc(method, tavg, tmax, tmin)
}
\arguments{
\item{method}{a character string indicating the method name to calculate PET. Available options are:
\itemize{
\item{"hs" for Hargreaves Samani method (tavg, tmax and tmin are required),}
\item{"od" for Oudin method (tavg required),}
\item{"mb" for McGuinness and Bordne method (tavg required),}
\item{"jh" for Jensen Haise method (tavg required),}
\item{"br" for Baier and Robertson (tmax and tmin are required).}
}}

\item{tavg}{a RasterBrick object having average temperature}

\item{tmax}{a RasterBrick object having maximum temperature}

\item{tmin}{a RasterBrick object having minimum temperature}
}
\value{
a RasterBrick object
}
\description{
The function \code{pet_calc} calculate selected PET method
}
\examples{
#Calculate PET by Hargreaves-Samani method 
download_terraclimate(folder_path = ".","land", variable = "t")   
tavg_brick <- raster::brick("terraclimate_tavg_land_19580101_20221231_025_monthly.nc")
tmax_brick <- raster::brick("terraclimate_tmax_land_19580101_20221231_025_monthly.nc")
tmin_brick <- raster::brick("terraclimate_tmin_land_19580101_20221231_025_monthly.nc")
pet_hs <- pet_calc(method = "hs",tavg = tavg_brick, tmax = tmax_brick, tmin = tmin_brick)
}
