% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet_calc.R
\name{pet_calc}
\alias{pet_calc}
\title{Calculate PET by various offline methods}
\usage{
pet_calc(method = "od", tavg = NULL, tmax = NULL, tmin = NULL)
}
\arguments{
\item{method}{a character string indicating the method name to calculate PET. Available options are:
\itemize{
\item{"hs" for Hargreaves Samani method,}
\item{"od" for Oudin method,}
\item{"mb" for McGuinness and Bordne method,}
\item{"jh" for Jensen Haise method,}
\item{"br" for Baier and Robertson.}
}}

\item{tavg}{a RasterBrick object having average temperature}

\item{tmax}{a RasterBrick object having maximum temperature}

\item{tmin}{a RasterBrick object having minimum temperature}
}
\value{
a RasterBrick object
}
\description{
The function \code{pet_calc} calculate selected PET method
}
\details{
Different PET methods require different meteorological inputs, such 
as tmax (maximum temperature), tmin (minimum temperature), and tavg (average temperature). 
The required inputs corresponding to each PET method are listed as follows:
\itemize{
\item{"hs" (Hargreaves-Samani): tavg, tmin, tmax,}
\item{"od" (Oudin): tavg,}
\item{"mb" (McGuinness and Bordne method): tavg,}
\item{"jh" (Jensen-Haise): tavg,}
\item{"br" (Baier and Robertson): tmin, tmax.}
}
}
\examples{
\dontrun{
#Calculate PET by Hargreaves-Samani method 
Function \code{\link{download_terraclimate}} is used to download temperature data
tavg_brick <- raster::brick("terraclimate_tavg_land_19580101_20221231_025_monthly.nc")
tmax_brick <- raster::brick("terraclimate_tmax_land_19580101_20221231_025_monthly.nc")
tmin_brick <- raster::brick("terraclimate_tmin_land_19580101_20221231_025_monthly.nc")
pet_hs <- pet_calc(method = "hs",tavg = tavg_brick, tmax = tmax_brick, tmin = tmin_brick)}
}
