% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet_hs.R
\name{pet_hs}
\alias{pet_hs}
\title{PET calculation by Hargreaves-Samani method}
\usage{
pet_hs(tavg, tmax, tmin)
}
\arguments{
\item{tavg}{a RasterBrick object having average temperature}

\item{tmax}{a RasterBrick object having maximum temperature}

\item{tmin}{a RasterBrick object having minimum temperature}
}
\value{
a RasterBrick object
}
\description{
Function to calculate PET
}
\examples{
function \code{download_terraclimate} is used to download TerraClimate temperature data (tmax,tmin and tavg)
tavg_brick <- raster::brick("terraclimate_tavg_land_19580101_20221231_025_monthly.nc")
tmax_brick <- raster::brick("terraclimate_tmax_land_19580101_20221231_025_monthly.nc")
tmin_brick <- raster::brick("terraclimate_tmin_land_19580101_20221231_025_monthly.nc")
pet_hs <- pet_calc(method = "hs",tavg_brick, tmax_brick, tmin_brick)
}
\keyword{internal}
