% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet_jh.R
\name{pet_jh}
\alias{pet_jh}
\title{PET calculation by Jensen Haise method}
\usage{
pet_jh(tavg)
}
\arguments{
\item{tavg}{a RasterBrick object having average temperature}
}
\value{
a RasterBrick object
}
\description{
Function to calculate pet
}
\examples{
\dontrun{
x <- tavg.nc  or character string with path to data file 
x_brick <- raster::brick(x) 
pet <- pet_jh(tavg = x_brick)
}
}
\keyword{internal}
