% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet_jh.R
\name{pet_jh}
\alias{pet_jh}
\title{PET calculation by Jensen Haise method}
\usage{
pet_jh(tavg)
}
\arguments{
\item{tavg}{a RasterBrick object having average temperature}
}
\value{
a RasterBrick object
}
\description{
Function to calculate pet
}
\examples{
download_terraclimate(folder_path = ".","land", variable = "tavg") 
tavg_brick <- raster::brick("terraclimate_tavg_land_19580101_20221231_025_monthly.nc")
pet_jh <- pet_calc(method = "jh", tavg = tavg_brick)
}
\keyword{internal}
