% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet_mb.R
\name{pet_mb}
\alias{pet_mb}
\title{PET calculation by McGuinness Bordne method}
\usage{
pet_mb(tavg)
}
\arguments{
\item{tavg}{a RasterBrick object having average temperature}
}
\value{
a RasterBrick object
}
\description{
Function to calculate pet getZ setZ
}
\examples{
#Function \code{\link{download_terraclimate}} is used to download TerraClimate temperature data (tavg) 
tavg_brick <- raster::brick("terraclimate_tavg_land_19580101_20221231_025_monthly.nc")
pet_mb <- pet_calc(method = "mb", tavg = tavg_brick)
}
\keyword{internal}
