% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet_mb.R
\name{pet_mb}
\alias{pet_mb}
\title{PET calculation by McGuinness Bordne method}
\usage{
pet_mb(tavg)
}
\arguments{
\item{tavg}{a RasterBrick object having average temperature}
}
\value{
a RasterBrick object
}
\description{
Function to calculate pet
}
\examples{
#Downloading temperature data
download_terraclimate(folder_path = ".","land", variable = "tavg") 
#Loading the nc files     
tavg_nc <- "terraclimate_tavg_land_19580101_20221231_025_monthly.nc"
# Converting nc files to RasterBrick 
tavg_brick <- raster::brick(tavg_nc)
#Calculating PET
pet_mb <- pet_calc(method_name = "mb", tavg = tavg_brick)
}
\keyword{internal}
