% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet_od.R
\name{pet_od}
\alias{pet_od}
\title{PET calculation by Oudin method}
\usage{
pet_od(tavg)
}
\arguments{
\item{tavg}{a RasterBrick object having average temperature}
}
\value{
a RasterBrick object
}
\description{
Function to calculate pet
}
\examples{
\dontrun{
x <- tavg.nc  or character string with path to data file 
x_brick <- raster::brick(x) 
pet <- pet_od(tavg = x_brick)
}
}
\keyword{internal}
