% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet_od.R
\name{pet_od}
\alias{pet_od}
\title{PET calculation by Oudin method}
\usage{
pet_od(tavg)
}
\arguments{
\item{tavg}{a RasterBrick object having average temperature}
}
\value{
a RasterBrick object
}
\description{
Function to calculate pet
}
\examples{
download_terraclimate(folder_path = ".","land", variable = "tavg") 
tavg_brick <- raster::brick("terraclimate_tavg_land_19580101_20221231_025_monthly.nc")
pet_od <- pet_calc(method = "od", tavg = tavg_brick)
}
\keyword{internal}
